Move the chat-input module in input directory
