+++ packages/jupyter-chat/src/components/input/index.ts
@@ -4,5 +4,6 @@
  */
 
 export * from './buttons';
+export * from './chat-input';
 export * from './toolbar-registry';
 export * from './use-chat-commands';
